Add tests and missing endpoint

+++ src/main/java/com/ml/mutant/domain/Stats.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class Stats {
 
-    private int countMutantDna;
+    private long countMutantDna;
 
-    private int countHumanDna;
+    private long countHumanDna;
 
     public double getRatio () {
         return (double) countMutantDna/countHumanDna;
